extend file tracker instead of storage
